+++ web/app/components-v2/opiskeluoikeus/OpiskeluoikeusTitle.tsx
@@ -1,36 +1,22 @@
-import React, { useCallback, useMemo, useRef, useState } from 'react'
-import {
-  createLocalThenApiCache,
-  isSuccess,
-  useApiWithParams
-} from '../../api-fetch'
+import React from 'react'
 import { TreeHook } from '../../appstate/tree'
-import { useVersionumero } from '../../appstate/useSearchParam'
-import { TestIdLayer, TestIdRoot, TestIdText } from '../../appstate/useTestId'
+import { TestIdRoot, TestIdText } from '../../appstate/useTestId'
 import { useKansalainenTaiSuoritusjako } from '../../appstate/user'
-import {
-  ISO2FinnishDateTime,
-  formatYearRange,
-  yearFromIsoDateString
-} from '../../date/date'
+import { formatYearRange, yearFromIsoDateString } from '../../date/date'
 import { t } from '../../i18n/i18n'
 import ChevronDownIcon from '../../icons/ChevronDownIcon'
 import ChevronUpIcon from '../../icons/ChevronUpIcon'
 import { Opiskeluoikeus } from '../../types/fi/oph/koski/schema/Opiskeluoikeus'
-import { last, nonNull } from '../../util/fp/arrays'
-import { fetchVersiohistoria } from '../../util/koskiApi'
+import { nonNull } from '../../util/fp/arrays'
 import { viimeisinOpiskelujaksonTila } from '../../util/schema'
 import { uncapitalize } from '../../util/strings'
-import { currentQueryWith, parseQuery } from '../../util/url'
-import { CommonProps, common, cx } from '../CommonProps'
+import { CommonProps, common } from '../CommonProps'
 import { VirkailijaOnly } from '../access/VirkailijaOnly'
 import { Column, ColumnRow } from '../containers/Columns'
-import { PositionalPopup } from '../containers/PositionalPopup'
 import { ExpandButton } from '../controls/ExpandButton'
-import { FlatButton } from '../controls/FlatButton'
-import { LinkButton } from '../controls/LinkButton'
 import { Lowercase } from '../texts/Lowercase'
 import { Trans } from '../texts/Trans'
+import { VersiohistoriaButton } from './VersiohistoriaButton'
 
 export type OpiskeluoikeusTitleProps = CommonProps<{
   opiskeluoikeus: Opiskeluoikeus
@@ -137,112 +123,4 @@ export const OpiskeluoikeusTitle = (props: OpiskeluoikeusTitleProps) => {
       )}
     </TestIdRoot>
   )
-}
-
-type VersiohistoriaButtonProps = CommonProps<{
-  opiskeluoikeusOid: string
-}>
-
-const VersiohistoriaButton: React.FC<VersiohistoriaButtonProps> = (props) => {
-  const buttonRef = useRef(null)
-  const [versiohistoriaVisible, setVersiohistoriaVisible] = useState(false)
-  const toggleList = useCallback(
-    () => setVersiohistoriaVisible(!versiohistoriaVisible),
-    [versiohistoriaVisible]
-  )
-  const hideList = useCallback(() => setVersiohistoriaVisible(false), [])
-  const currentVersion = useVersionumero()
-
-  return (
-    <TestIdLayer id="versiohistoria">
-      <span className="VersiohistoriaButton" ref={buttonRef}>
-        <FlatButton
-          onClick={toggleList}
-          aria-haspopup="menu"
-          aria-expanded={versiohistoriaVisible}
-          testId="button"
-        >
-          {currentVersion
-            ? `${t('Versionumero')}: v${currentVersion}`
-            : t('Versiohistoria')}
-        </FlatButton>
-        <PositionalPopup
-          align="right"
-          onDismiss={hideList}
-          open={versiohistoriaVisible}
-          parentRef={buttonRef}
-        >
-          <VersiohistoriaList
-            opiskeluoikeusOid={props.opiskeluoikeusOid}
-            open={versiohistoriaVisible}
-          />
-        </PositionalPopup>
-      </span>
-    </TestIdLayer>
-  )
-}
-
-type VersiohistoriaListProps = CommonProps<{
-  opiskeluoikeusOid: string
-  open: boolean
-}>
-
-const versiolistaCache = createLocalThenApiCache(fetchVersiohistoria)
-
-const VersiohistoriaList: React.FC<VersiohistoriaListProps> = (props) => {
-  const historia = useApiWithParams(
-    fetchVersiohistoria,
-    props.open ? [props.opiskeluoikeusOid] : undefined,
-    versiolistaCache
-  )
-
-  const versioParam = useVersionumero()
-
-  const currentVersion = useMemo(() => {
-    return versioParam
-      ? parseInt(versioParam)
-      : isSuccess(historia)
-        ? last(historia.data)?.versionumero
-        : undefined
-  }, [historia, versioParam])
-
-  return isSuccess(historia) ? (
-    <TestIdLayer id="list">
-      <ul className="VersiohistoriaList" role="navigation">
-        {historia.data.map((versio) => (
-          <li
-            key={versio.versionumero}
-            className={cx(
-              'VersiohistoriaList__item',
-              currentVersion === versio.versionumero &&
-                'VersiohistoriaList__item--current'
-            )}
-          >
-            <LinkButton
-              href={currentQueryWith({
-                opiskeluoikeus: props.opiskeluoikeusOid,
-                versionumero: versio.versionumero
-              })}
-              testId={versio.versionumero}
-            >
-              {`v${versio.versionumero}`}{' '}
-              {ISO2FinnishDateTime(versio.aikaleima)}
-            </LinkButton>
-          </li>
-        ))}
-        {versioParam && (
-          <li className="VersiohistoriaList__item">
-            <LinkButton
-              href={currentQueryWith({
-                opiskeluoikeus: props.opiskeluoikeusOid,
-                versionumero: null
-              })}
-            >
-              {t('Poistu versiohistoriasta')}
-            </LinkButton>
-          </li>
-        )}
-      </ul>
-    </TestIdLayer>
-  ) : null
 }
